Update policy.d spec to reflect what actually happened
These changes are not about 'intent', but rather how the intent
translated into practice.  There were two charms that didn't need
overrides, and this created space for an additional charm to be included
in the 19.10 release cycle.  Also, more in-depth testing was possible,
and the spec is updated to reflect that.

Change-Id: Idcd59828ce81581f8e088948a3d55b38b06cbdf7