Disable vTPM capabilities by default
Disable the vTPM capabilities by default. The necessary dependencies
have not been backported to focal (as of yet), requiring the packages to
be provided in a PPA. Rather than install this PPA by default, the charm
will disable the vTPM features by default and require the user to enable
them.

Change-Id: Id0521b4dd6a3e5a768cab14855e4f3fec240ff77